+++ app_reactnative/README.md
@@ -1,0 +1,5 @@
+## IPアドレスの変更
+
+## gitignoreされているファイルの追加
+'google-services.json'  
+ExpoでAndroid端末に通知を送るためにfirebaseを使用．その関連ファイル．
