+++ app_reactnative/gameanhour/App.js
@@ -4,8 +4,7 @@ import { View, Container, Header, Title, Content, Footer, FooterTab, Button, Lef
 import * as Font from 'expo-font';
 import { Ionicons } from '@expo/vector-icons';
 import { Mainscreen } from './components/page1/Mainscreen.js';
-import { GetMACAddr } from './components/page1/test.js';
-import { Notifications } from 'expo';
+import { GetMACAddr } from './components/page1/GetMAC.js';
 import * as Permissions from 'expo-permissions';
 import { createDrawerNavigator, createStackNavigator, createBottomTabNavigator, createAppContainer } from 'react-navigation';
 
@@ -30,7 +29,6 @@ export default class App extends React.Component {
   }
 
   async componentDidMount() {
-    registerForPushNotificationsAsync();
     await Font.loadAsync({
       Roboto: require('native-base/Fonts/Roboto.ttf'),
       Roboto_medium: require('native-base/Fonts/Roboto_medium.ttf'),
@@ -46,30 +44,4 @@ export default class App extends React.Component {
   
     return <AppContainer />;
   }
-}
-
-
-async function registerForPushNotificationsAsync() {
-  const { status: existingStatus } = await Permissions.getAsync(
-    Permissions.NOTIFICATIONS
-  );
-  let finalStatus = existingStatus;
-
-  // only ask if permissions have not already been determined, because
-  // iOS won't necessarily prompt the user a second time.
-  if (existingStatus !== 'granted') {
-    // Android remote notification permissions are granted during the app
-    // install, so this will only ask on iOS
-    const { status } = await Permissions.askAsync(Permissions.NOTIFICATIONS);
-    finalStatus = status;
-  }
-
-  // Stop here if the user did not grant permissions
-  if (finalStatus !== 'granted') {
-    return;
-  }
-
-  // Get the token that uniquely identifies this device
-  let token = await Notifications.getExpoPushTokenAsync();
-  console.log(token);
 }
